Separated BangBuilder from BangBus class (#58)
